# Symbols: run the faster regex first