@chrismwendt 2019-02-14
In the future we might want to list symbols in the current file, which would benefit from the file filter being run first.